requirements.txt
jinja2
feedparser
pyyaml
markdown
requests